2025-07-01-ai-content-pipeline-2/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI Content Pipeline Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.13",
    "pydantic>=2.11.7",
    "uvicorn[standard]>=0.32.1",
    "python-multipart>=0.0.20",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.1",
    "supabase>=2.10.0",
    "google-auth>=2.30.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.130.0",
    "baml-py==0.90.2",
    "requests>=2.31.0",
    "supersonic>=0.1.0",
    "cased-kit>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "isort>=5.13.0",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "ruff>=0.12.0",
]
